+++ src/csharp/Microsoft.Spark/ML/Feature/Base.cs
@@ -2,15 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using Microsoft.Spark;
 using Microsoft.Spark.Sql;
 using Microsoft.Spark.Sql.Types;
-using Microsoft.Spark.ML.Feature;
-using Microsoft.Spark.ML.Feature.Param;
 using Microsoft.Spark.Interop;
 using Microsoft.Spark.Interop.Ipc;
