+++ src/csharp/Microsoft.Spark/ML/Util/Write.cs
@@ -74,60 +74,4 @@ namespace Microsoft.Spark.ML.Feature
         /// <param name="path">The path to save the object to</param>
         void Save(string path);
     }
-
-    /// <summary>
-    /// Class for utility classes that can load ML instances.
-    /// </summary>
-    /// <typeparam name="T">ML instance type</typeparam>
-    public class ScalaMLReader<T> : IJvmObjectReferenceProvider
-    {
-        public ScalaMLReader(JvmObjectReference jvmObject) => Reference = jvmObject;
-
-        public JvmObjectReference Reference { get; private set; }
-
-        /// <summary>
-        /// Loads the ML component from the input path.
-        /// </summary>
-        /// <param name="path">The path the previous instance of type T was saved to</param>
-        /// <returns>The type T instance</returns>
-        public T Load(string path) =>
-            WrapAsType((JvmObjectReference)Reference.Invoke("load", path));
-
-        /// <summary>Sets the Spark Session to use for saving/loading.</summary>
-        /// <param name="sparkSession">The Spark Session to be set</param>
-        public ScalaMLReader<T> Session(SparkSession sparkSession)
-        {
-            Reference.Invoke("session", sparkSession);
-            return this;
-        }
-
-        private static T WrapAsType(JvmObjectReference reference)
-        {
-            ConstructorInfo constructor = typeof(T)
-                .GetConstructors(BindingFlags.NonPublic | BindingFlags.Instance)
-                .Single(c =>
-                {
-                    ParameterInfo[] parameters = c.GetParameters();
-                    return (parameters.Length == 1) &&
-                        (parameters[0].ParameterType == typeof(JvmObjectReference));
-                });
-
-            return (T)constructor.Invoke(new object[] { reference });
-        }
-    }
-
-    /// <summary>
-    /// Interface for objects that provide MLReader.
-    /// </summary>
-    /// <typeparam name="T">
-    /// ML instance type
-    /// </typeparam>
-    public interface ScalaMLReadable<T>
-    {
-        /// <summary>
-        /// Get the corresponding ScalaMLReader instance.
-        /// </summary>
-        /// <returns>an <see cref="ScalaMLReader&lt;T&gt;"/> instance for this ML instance.</returns>
-        ScalaMLReader<T> Read();
-    }
 }
